+++ sdk/go/kubernetes/core/v1/secretPatch.go
@@ -57,6 +57,17 @@ func NewSecretPatch(ctx *pulumi.Context,
 
 	args.ApiVersion = pulumi.StringPtr("v1")
 	args.Kind = pulumi.StringPtr("Secret")
+	if args.Data != nil {
+		args.Data = pulumi.ToSecret(args.Data).(pulumi.StringMapInput)
+	}
+	if args.StringData != nil {
+		args.StringData = pulumi.ToSecret(args.StringData).(pulumi.StringMapInput)
+	}
+	secrets := pulumi.AdditionalSecretOutputs([]string{
+		"data",
+		"stringData",
+	})
+	opts = append(opts, secrets)
 	opts = utilities.PkgResourceDefaultOpts(opts)
 	var resource SecretPatch
 	err := ctx.RegisterResource("kubernetes:core/v1:SecretPatch", name, args, &resource, opts...)