+++ sdk/java/src/main/java/com/pulumi/kubernetes/core/v1/SecretPatch.java
@@ -12,6 +12,7 @@ import com.pulumi.kubernetes.core.v1.SecretPatchArgs;
 import com.pulumi.kubernetes.meta.v1.outputs.ObjectMetaPatch;
 import java.lang.Boolean;
 import java.lang.String;
+import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import javax.annotation.Nullable;
@@ -180,6 +181,10 @@ public class SecretPatch extends com.pulumi.resources.CustomResource {
     private static com.pulumi.resources.CustomResourceOptions makeResourceOptions(@Nullable com.pulumi.resources.CustomResourceOptions options, @Nullable Output<java.lang.String> id) {
         var defaultOptions = com.pulumi.resources.CustomResourceOptions.builder()
             .version(Utilities.getVersion())
+            .additionalSecretOutputs(List.of(
+                "data",
+                "stringData"
+            ))
             .build();
         return com.pulumi.resources.CustomResourceOptions.merge(defaultOptions, options, id);
     }